S3-BACKEND-NOTES : modify structure to use mongoDb from docker image and H2 DB for test + different profile set, for test, for local use and for docker use

+++ backend-note/src/main/java/com/mediLaboSolutions/backendnote/configData/InitData.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mediLaboSolutions.backendnote.models.Note;
 import com.mediLaboSolutions.backendnote.repositories.NoteRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
@@ -15,6 +16,9 @@ import java.util.List;
 @Component
 public class InitData implements CommandLineRunner {
 
+    @Value("${spring.profiles.active:defaultProfile}")
+    private String activeProfile;
+
     private final NoteRepository noteRepository;
 
     public InitData(NoteRepository noteRepository) {
@@ -23,24 +27,30 @@ public class InitData implements CommandLineRunner {
 
     @Override
     public void run(String... args) throws Exception {
-        // Load data from json file
-        ObjectMapper objectMapper = new ObjectMapper();
-        TypeReference<List<Note>> typeReference = new TypeReference<>() {
-        };
 
-        InputStream inputStream = TypeReference.class.getResourceAsStream("/notesData.json");
+        if (activeProfile.contains("test")) {
 
-        try {
-            //Delete all present data
-            noteRepository.deleteAll();
+            log.info("|||||||||||||  Initialisation des données désactivé pour le profil 'test'.  |||||||||||||");
+        } else {
 
-            //Insert new data
-            List<Note> entities = objectMapper.readValue(inputStream, typeReference);
-            noteRepository.saveAll(entities);
-            log.info("|||||||||||||  Initialisation des données réussie !  |||||||||||||");
-        } catch (Exception e) {
-            log.info("|||||||||||||  Erreur lors de l'initialisation des données : " + e.getMessage() + "  |||||||||||||");
+            TypeReference<List<Note>> typeReference = new TypeReference<>() {};
+            ObjectMapper objectMapper = new ObjectMapper();
+
+            InputStream inputStream = TypeReference.class.getResourceAsStream("/notesData.json");
+
+            try {
+                noteRepository.deleteAll();
+
+                List<Note> entities = objectMapper.readValue(inputStream, typeReference);
+
+                noteRepository.saveAll(entities);
+                log.info("|||||||||||||  Initialisation des données réussie!  |||||||||||||");
+
+            } catch (Exception e) {
+                log.error("|||||||||||||  Erreur lors de l'initialisation des données : " + e.getMessage() + "  |||||||||||||");
+            }
         }
-
     }
 }
+
+
